% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fAlldata}
\alias{fAlldata}
\title{Return ff data based on params table}
\usage{
fAlldata(params)
}
\arguments{
\item{params}{table with layers' characeristics}
}
\value{
a ff data table
}
\description{
Return ff data based on params table
}
