% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_related_functions.R
\name{get_cluster_locations}
\alias{get_cluster_locations}
\title{Title}
\usage{
get_cluster_locations(mytree, leaves = FALSE)
}
\arguments{
\item{mytree}{a classification tree}

\item{leaves}{if TRUE, also returns the location of the 2 leaves}
}
\description{
Title
}
