% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_related_functions.R
\name{get_clusters_containing_individual}
\alias{get_clusters_containing_individual}
\title{Title}
\usage{
get_clusters_containing_individual(mytree, i)
}
\arguments{
\item{i}{}
}
\description{
Title
}
