% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_related_functions.R
\name{plot_tree}
\alias{plot_tree}
\title{Title}
\usage{
plot_tree(mytree, ylim = c(0, 1.2 * max(mytree$height)), labels_ab = FALSE,
  nclust)
}
\arguments{
\item{nclust}{}
}
\description{
Title
}
