% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_analyses.R
\name{dudi.acm}
\alias{dudi.acm}
\title{This is a slightly modified version of the dudi.acm function, that deals with columns
with null weight}
\usage{
dudi.acm(df, row.w = rep(1, nrow(df)), scannf = TRUE, nf = 2)
}
\arguments{
\item{nf}{}
}
\value{
an MCA result
}
\description{
This is a slightly modified version of the dudi.acm function, that deals with columns
with null weight
}
