% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_related_functions.R
\name{get_lower_clusters}
\alias{get_lower_clusters}
\title{Get clusters ID or individuals' indices, of clusters or individuals comprised in a cluster}
\usage{
get_lower_clusters(mytree, i, immediate = FALSE)
}
\arguments{
\item{mytree}{a classification tree}

\item{i}{the index of the cluster}

\item{immediate}{if TRUE, the function returns clusters at the level immediately below, if FALSE, the function returns all individuals comprised in the cluster}
}
\value{
vector of indices
}
\description{
Get clusters ID or individuals' indices, of clusters or individuals comprised in a cluster
}
\examples{
hc <- hclust(dist(USArrests), "ave")
get_lower_clusters(hc,15,immediate=TRUE)
get_lower_clusters(hc,15,immediate=FALSE)
}
